Allow in progress recording downloads (#21)
### Description

We also want to be able to show recordings of the session so far. The way we've hooked up `ffmpeg` allows this already! The work in this pr is to:

1. Relax constraints around downloading in progress recordings
2. Instead, we'll check the file size and return a `202` if the recording doesn't have a reasonable amount of data. For the moment we've picked 100 bytes, which is substantially smaller than what `ffmpeg` will flush to disk

Also lurked in some minor config validation + cleaned up the `Start` logic as recommended by my cursor agents

### Testing

Via the local docker set up:

1. Build and run the container with the recording server
2. Start recording
3. Immediately hit the download -> confirm a `202` response
4. (wait for a bit)
5. Hit the download again -> confirm valid video output
6. Stop recording
7. Download -> confirm valid video output