Block scale to zero during recording, allow multiple recordings (#29)
### Description

As we've stress tested the recording server more we've noticed the behavior from pause <> resume to be a bit more unpredictable than we'd like. Instead we're opt'ing to leverage the pseudo file detailed in [the unikraft docs](https://unikraft.cloud/docs/api/v1/instances/#scale-to-zero) to disable scale to zero while we're recording.


- The interface doesn't change - the expectation is that the caller will still call `Stop` but the browser instance will stay live + recording for the duration between `[Start, Stop]`
- I've relaxed the condition of having only 1 active recorder alive. I've tested `ffmpeg` is happy with this in both runtimes. Note: we expect names to be unique and we will reject any duplicates.
- I've also added a timeout that we thread into `ffmpeg` that will work in conjunction with the existing max size as a guardrail.
- Added a method to list all recorders for simpler integration + export

The main thing to call out here is I've pushed the implementation of the scale to zero down into the recorder so it will be re-enabled as soon as possible (e.g. the moment the recording is complete or the user has explicitly called to `Stop` recording. These may race which is why I've wrapped this in a sync.Once to make it more forgiving)

### Testing

Tested with the end to end automation suite. The expected behavior that the instance would stay in `running` until shutdown was explicitly called was confirmed